Add workaround for missing UUID parameter for data source
