--- nutanix/data_source_address_group.go
@@ -1,40 +1,0 @@
-package nutanix
-
-import "github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-
-func dataSourceNutanixAddressGroup() *schema.Resource {
-	return &schema.Resource{
-		Read: resourceNutanixAddressGroupRead,
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Required: true,
-			},
-			"description": {
-				Type:     schema.TypeString,
-				Optional: true,
-			},
-			"ip_address_block_list": {
-				Type:     schema.TypeList,
-				Required: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"ip": {
-							Type:     schema.TypeString,
-							Required: true,
-						},
-						"prefix_length": {
-							Type:     schema.TypeInt,
-							Required: true,
-						},
-					},
-				},
-			},
-			"address_group_string": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-		},
-	}
-
-}
